# CNDB-12382 Implement Counter locks without Striped locks