… src/java/org/apache/cassandra/db/counters/CounterLockManager.java
… src/java/org/apache/cassandra/db/counters/StripedCounterLockManager.java
… test/unit/org/apache/cassandra/db/counters/CounterLockManagerTest.java
… test/unit/org/apache/cassandra/metrics/CounterMutationMetricsTest.java