### What is the issue
While working on CNDB-12382 I have seen that increasing the number of stripes helps to solve the timeout problems.

### What does this PR fix and why was it fixed
This PR introduces an abstraction over the LOCKS mechanism in CounterMutation and provides a different implementation.

The new implementation is not based on striping the locks (that is about sharing the same lock for different keys) but it maintains a set of reference counted objects that wrap the actual locks.

(The implementation of the Lock manager is the adaptaction of a similar mechanism taken from another OSS project)

This implementation also makes it clearer that the key for the locks is an Integer, computed  by summing some hash codes, and this already creates some kind of bucketing.
It is still possible that two mutations fall into the same lock key.

When do the new implementation works better ?
- it should generally works better for most of the workloads
- when there are few active counter mutations (the number of locks is usually expected to be like the size of the Counters Mutation Stage thread pool)

Rollback:
Add -Dcassandra.use_striped_counter_lock_manager=true to the writers JVM options


Licensing: this code is copied and adapted  from
[LocalLockManager](https://github.com/diennea/herddb/blob/master/herddb-utils/src/main/java/herddb/utils/LocalLockManager.java) from the HerdDB project (Apache 2 licensed).