@eolivelli 2025-02-25
I have run the Astra Certification workload with this patch, the results are very good

previous code:
```
Test Duration              : 6h 0m 16s
Total Operations           : 4319555
Op Rate                    : 199 op/sec
Min Latency                : 8.195 ms
Avg Latency                : 14.094 ms
Median Latency             : 11.777 ms
95th Latency               : 19.385 ms
99th Latency               : 65.317 ms
99.9th Latency             : 162.562 ms
Max Latency                : 2052.325 ms
Median Absolute Deviation  : 0.808 ms
Interquartile Range        : 1.734 ms
```
new code
```
Test Duration              : 6h 0m 16s
Total Operations           : 4319999
Op Rate                    : 199 op/sec
Min Latency                : 8.220 ms
Avg Latency                : 13.286 ms
Median Latency             : 11.848 ms
95th Latency               : 15.136 ms
99th Latency               : 53.305 ms
99.9th Latency             : 133.009 ms
Max Latency                : 843.284 ms
Median Absolute Deviation  : 0.782 ms
Interquartile Range        : 1.588 ms
```


Counter acquire lock timeout metrics in Astra (on the left you can see the new code, on the right the old code)

![image](https://github.com/user-attachments/assets/0efa3e7f-99f0-42b9-9b4a-397c3c145a6f)