@JeremiahDJordan 2025-02-26
> The implementation of the Lock manager is the adaptaction of a similar mechanism taken from another OSS project
Can we add attribution to the project this came from, and probably even a link to code in case we have issues and want to check for bug fixes in the original.