@eolivelli 2025-02-27
The original implementation is here
https://github.com/diennea/herddb/blob/master/herddb-utils/src/main/java/herddb/utils/LocalLockManager.java

 There is no need to mention it in a comment on the code, the code is pretty straightforward, it was code that I wrote myself long time ago and actaully the implemetation here diverged a little bit, because we need less complexity